# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Philippine Schools API",
    "description": "Read-only API for searching and filtering Philippine school records.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/schools": {
      "get": {
        "tags": [
          "Schools"
        ],
        "summary": "List schools",
        "description": "Returns paginated schools with optional text search and filters.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search term for id, name, head, address, region",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "description": "Case-insensitive partial region match",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region_id",
            "in": "query",
            "description": "Exact region id match",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Exact school type match",
            "schema": {
              "type": "string",
              "example": "elementary_school"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Exact category match",
            "schema": {
              "type": "string",
              "example": "public"
            }
          },
          {
            "name": "curricular_classification",
            "in": "query",
            "description": "Partial curricular classification match",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 25
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "id",
                "name",
                "region",
                "type",
                "category"
              ],
              "default": "name"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated schools",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/School"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/schools/{id}": {
      "get": {
        "tags": [
          "Schools"
        ],
        "summary": "Get school by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "School found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/School"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "School not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total_pages": {
            "type": "integer"
          },
          "has_next_page": {
            "type": "boolean"
          },
          "has_prev_page": {
            "type": "boolean"
          }
        }
      },
      "School": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "436005"
          },
          "name": {
            "type": "string",
            "example": "\"Eskwela\" Phonics Reading Center"
          },
          "head": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "curricular_classification": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "region_id": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}